% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_labels.R
\name{find_previous_actions}
\alias{find_previous_actions}
\title{Does the indexing to find the goals/concessions in a df of spadl events}
\usage{
find_previous_actions(action_id, team, n_prev_actions, spadl)
}
\arguments{
\item{action_id}{The action_id of an action}

\item{team}{The team_id of an action}

\item{n_prev_actions}{The number of events to look backwards and class as 'leading to a goal'}

\item{spadl}{A dataframe of event data in spadl format}
}
\description{
Does the indexing to find the goals/concessions in a df of spadl events
}
\author{
Robert Hickman
}
