% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_features.R
\name{get_context_goals}
\alias{get_context_goals}
\title{Get the goal context of a spadl data frame}
\usage{
get_context_goals(spadl)
}
\arguments{
\item{spadl}{A dataframe of event data in spadl format}
}
\description{
Get the goal context of a spadl data frame
}
\author{
Robert Hickman
}
