% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_config.R
\docType{data}
\name{pen_score_frac}
\alias{pen_score_frac}
\title{The fixed probability of scoring a penalty}
\format{
An object of class \code{numeric} of length 1.
}
\source{
\href{https://github.com/ML-KULeuven/socceraction/blob/master/socceraction/spadl/config.py}{socceraction repo}
}
\usage{
data(pen_score_frac)
}
\description{
The fixed probability of scoring a penalty
}
\examples{
score_chance <- pen_score_frac
}
\keyword{datasets}
