% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch_plot.R
\name{plot_pitch}
\alias{plot_pitch}
\title{Plot standard pitch with dimension of 105m x 68m (or yard equivalent).
Event or tracking data the has been standardized to these dimensions
can be readily plotted on top.}
\usage{
plot_pitch(
  unit = "meters",
  theme = "light",
  type = "full_pitch",
  orientation = "horizontal",
  direction_arrow = FALSE
)
}
\arguments{
\item{unit}{Dimensions in meters or yards. Defaults to meters}

\item{theme}{Choose between light and dark theme. Defaults to light}

\item{type}{Choose between full_pitch, offensive_half or defensive_half. Defaults to full_pitch}

\item{orientation}{Choose between horizontal or vertical orientation. Defaults to horizontal}

\item{direction_arrow}{Optional indicator for direction of play. Defaults to false}
}
\description{
Plot standard pitch with dimension of 105m x 68m (or yard equivalent).
Event or tracking data the has been standardized to these dimensions
can be readily plotted on top.
}
\author{
Lars Maurath
}
