% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sb_clean_locations}
\alias{sb_clean_locations}
\title{Clean location columns from import of StatsBomb event data}
\usage{
sb_clean_locations(match_events)
}
\arguments{
\item{match_events}{A dataframe of imported StatsBomb event data}
}
\description{
Clean location columns from import of StatsBomb event data
}
\examples{

cl_final <- data.frame(match_id = 22912)
sb_data <- StatsBombR::get.matchFree(cl_final)
sb_data_cleanlocs <- sb_clean_locations(sb_data)

}
\author{
Robert Hickman
}
