% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl.R
\name{sb_convert_spadl}
\alias{sb_convert_spadl}
\title{Convert a dataframe of StatsBomb data into the spadl format}
\usage{
sb_convert_spadl(match_events, save_dir = ".")
}
\arguments{
\item{match_events}{A dataframe of imported StatsBomb event data}
}
\description{
Convert a dataframe of StatsBomb data into the spadl format
}
\examples{

cl_final <- data.frame(match_id = 22912)
sb_data <- StatsBombR::get.matchFree(cl_final)
sb_spadl <- sb_convert_spadl(match_events)

}
\author{
Robert Hickman
}
