% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sb_getmins_played}
\alias{sb_getmins_played}
\title{Find minutes played and (guesstimate) position for players from Statsbomb event data}
\usage{
sb_getmins_played(match_events)
}
\arguments{
\item{match_events}{A dataframe of imported StatsBomb event data}
}
\description{
Find minutes played and (guesstimate) position for players from Statsbomb event data
}
\examples{

cl_final <- data.frame(match_id = 22912)
sb_data <- StatsBombR::get.matchFree(cl_final)
sb_data_cleanlocs <- sb_getmins_played(sb_data)

}
\author{
Robert Hickman
}
