% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_config.R
\docType{data}
\name{spadl_bodypart_ids}
\alias{spadl_bodypart_ids}
\title{Bodypart ids for spadl actions}
\format{
An object of class \code{data.frame} with 3 rows and 2 columns.
}
\source{
\href{https://github.com/ML-KULeuven/socceraction/blob/master/socceraction/spadl/config.py}{socceraction repo}
}
\usage{
data(spadl_bodypart_ids)
}
\description{
Bodypart ids for spadl actions
}
\examples{
bodyparts <- data(spadl_bodyparts_ids)
}
\keyword{datasets}
