% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl.R
\name{spadl_dict}
\alias{spadl_dict}
\title{Convert data from provider specs into a common spadl format}
\usage{
spadl_dict(type, provider, data)
}
\arguments{
\item{type}{A type of data to be provided}

\item{provider}{The provider of the event data}

\item{data}{A vector of event data to be converted}
}
\description{
Convert data from provider specs into a common spadl format
}
\examples{

}
\author{
Robert Hickman
}
