% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_config.R
\docType{data}
\name{spadl_field_length}
\alias{spadl_field_length}
\title{The field length for spadl actions}
\format{
An object of class \code{numeric} of length 1.
}
\source{
\href{https://github.com/ML-KULeuven/socceraction/blob/master/socceraction/spadl/config.py}{socceraction repo}
}
\usage{
data(spadl_field_length)
}
\description{
The field length for spadl actions
}
\examples{
results <- data(spadl_field_length)s
}
\keyword{datasets}
