% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_config.R
\docType{data}
\name{spadl_result_ids}
\alias{spadl_result_ids}
\title{Results ids for spadl actions}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\source{
\href{https://github.com/ML-KULeuven/socceraction/blob/master/socceraction/spadl/config.py}{socceraction repo}
}
\usage{
data(spadl_result_ids)
}
\description{
Results ids for spadl actions
}
\examples{
results <- data(spadl_result_ids)
}
\keyword{datasets}
