% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl_config.R
\docType{data}
\name{spadl_type_ids}
\alias{spadl_type_ids}
\title{Type ids for spadl actions}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://github.com/ML-KULeuven/socceraction/blob/master/socceraction/spadl/config.py}{socceraction repo}
}
\usage{
data(spadl_type_ids)
}
\description{
Type ids for spadl actions
}
\examples{
types <- data(spadl_type_ids)
}
\keyword{datasets}
