% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spadl.R
\name{split_dribbles}
\alias{split_dribbles}
\title{Add missing dribbles back into spadl data frame}
\usage{
split_dribbles(spadl)
}
\arguments{
\item{spadl}{A data frame of the prospective spadl format}
}
\description{
Add missing dribbles back into spadl data frame
}
\examples{

}
\author{
Robert Hickman
}
