% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_coordinates.R
\name{standardize_opta_x}
\alias{standardize_opta_x}
\title{Standardize length (x) location columns from Opta event data to 105m x 68m convention}
\usage{
standardize_opta_x(
  data,
  cols = c("location_x", "PassEndX", "BlockedX"),
  unit = "meters"
)
}
\arguments{
\item{data}{A dataframe of Opta event data}

\item{cols}{Columns to convert. Defaults to all x location columns}

\item{unit}{Convert to meters or yards. Defaults to meters}
}
\description{
Standardize length (x) location columns from Opta event data to 105m x 68m convention
}
\author{
Lars Maurath
}
