% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_coordinates.R
\name{standardize_opta_z}
\alias{standardize_opta_z}
\title{Standardize height (z) location columns from Opta event data to 105m x 68m convention}
\usage{
standardize_opta_z(data, cols = c("GoalMouthZ"), unit = "meters")
}
\arguments{
\item{data}{A dataframe of Opta event data}

\item{cols}{Columns to convert. Defaults to all z location columns}

\item{unit}{Convert to meters or yards. Defaults to meters}
}
\description{
Standardize height (z) location columns from Opta event data to 105m x 68m convention
}
\author{
Lars Maurath
}
