% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_coordinates.R
\name{standardize_statsbomb_x}
\alias{standardize_statsbomb_x}
\title{Standardize length (x) location columns from StatsBomb event data to 105m x 68m convention}
\usage{
standardize_statsbomb_x(
  data,
  cols = c("location.x", "carry.end_location.x", "pass.end_location.x",
    "shot.end_location.x", "location.x.GK"),
  unit = "meters"
)
}
\arguments{
\item{data}{A dataframe of cleaned StatsBomb event data}

\item{cols}{Columns to convert. Defaults to all x location columns}

\item{unit}{Convert to meters or yards. Defaults to meters}
}
\description{
Standardize length (x) location columns from StatsBomb event data to 105m x 68m convention
}
\examples{

cl_final <- data.frame(match_id = 22912)
sb_data <- StatsBombR::get.matchFree(cl_final)
sb_data <- allclean(sb_data)
sb_data <- standardize_statsbomb_x(sb_data)

}
\author{
Lars Maurath
}
