% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_coordinates.R
\name{standardize_statsbomb_y}
\alias{standardize_statsbomb_y}
\title{Standardize width (y) location columns from StatsBomb event data to 105m x 68m convention}
\usage{
standardize_statsbomb_y(
  data,
  cols = c("location.y", "carry.end_location.y", "pass.end_location.y",
    "shot.end_location.y", "location.y.GK"),
  unit = "meters"
)
}
\arguments{
\item{data}{A dataframe of cleaned StatsBomb event data}

\item{cols}{Columns to convert. Defaults to all y location columns}

\item{unit}{Convert to meters or yards. Defaults to meters}
}
\description{
Standardize width (y) location columns from StatsBomb event data to 105m x 68m convention
}
\examples{

cl_final <- data.frame(match_id = 22912)
sb_data <- StatsBombR::get.matchFree(cl_final)
sb_data <- allclean(sb_data)
sb_data <- standardize_statsbomb_y(sb_data)

}
\author{
Lars Maurath
}
