% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_features.R
\name{vaep_features_lag}
\alias{vaep_features_lag}
\title{Get a dataframe of n lagged columns of a feature (to be deprecated)}
\usage{
vaep_features_lag(vec, lags, name)
}
\arguments{
\item{vec}{A vector of data}

\item{lags}{An integer >= 0}

\item{name}{A character string}
}
\description{
Get a dataframe of n lagged columns of a feature (to be deprecated)
}
\examples{
feature <- runif(10)
lagged_cols <- vaep_features_lag(feature, 3, "lagged_feature_")
head(lagged_cols)

}
\author{
Robert Hickman
}
