% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_features.R
\name{vaep_get_features}
\alias{vaep_get_features}
\title{Get the features of each match event (and preceeding events) from a spadl format}
\usage{
vaep_get_features(spadl)
}
\arguments{
\item{spadl}{A dataframe of event data in spadl format}
}
\description{
Get the features of each match event (and preceeding events) from a spadl format
}
\author{
Robert Hickman
}
