% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_labels.R
\name{vaep_get_labels}
\alias{vaep_get_labels}
\title{Finds and labels goals in the next n actions from a df of spadl events}
\usage{
vaep_get_labels(spadl, n_prev_actions = 10)
}
\arguments{
\item{spadl}{A dataframe of event data in spadl format}

\item{n_prev_actions}{The number of events to look backwards and class as 'leading to a goal'}
}
\description{
Finds and labels goals in the next n actions from a df of spadl events
}
\author{
Robert Hickman
}
