% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_score.R
\name{vaep_get_scores}
\alias{vaep_get_scores}
\title{Calculate the VAEP score for all actions in a spadl df}
\usage{
vaep_get_scores(spadl, scores, concedes)
}
\arguments{
\item{spadl}{A df of spadl events}

\item{scores}{The name of a col of values calculated for the chance of scoring in next n actions}

\item{concedes}{The name of a col of values calculated for the chance of conceeding in next n actions}
}
\description{
Calculate the VAEP score for all actions in a spadl df
}
\author{
Robert Hickman
}
