% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaep_score.R
\name{vaep_score_actions}
\alias{vaep_score_actions}
\title{Calculate the VAEP score for all actions in a spadl df}
\usage{
vaep_score_actions(spadl, scores, concedes, score_type)
}
\arguments{
\item{spadl}{A df of spadl events}

\item{scores}{The name of a col of values calculated for the chance of scoring in next n actions}

\item{concedes}{The name of a col of values calculated for the chance of conceeding in next n actions}

\item{score_type}{Whether to score for the attacking contribution or defensive contribution}
}
\description{
Calculate the VAEP score for all actions in a spadl df
}
\author{
Robert Hickman
}
